Move query params tests to new TAP format; fixes botched merge
When rebasing this branch I missed the changes made to the tests
and so ended up with a bad merge. This commit fixes it.
